Returns the name of the cache file associated with the given URL. Note that this file may not actually exist, this function just provides the mapping from URL to filename

Returns the name of the cache file associated with the given URL. Note that this file may not actually exist, this function just provides the mapping from URL to filename

ala_cache_filename(url)

Arguments

url
string: the URL

Value

string: the file path and name

References

http://api.ala.org.au/

See also

Examples

ala_cache_filename("http://biocache.ala.org.au/ws/index/fields")

#> [1] "/var/folders/x5/g4flwxsd17z0b8vxsj8kcpyxmnwjtg/T//Rtmp0QGzs1/b2fa59b0f3a1de10b06a17d8a1616059"