List species matching criteria

List species matching criteria

ala_checklist(taxon_id, filters, geometry, caching = FALSE)

Arguments

taxon_id
string: single species ID or vector of species ids. Use `ala_taxa()` to get lookup species id.

filters
string: a list to narrow down the search, in the form `list(field = value)`. To limit the query to a year range, use `list(year = c(1990, 2000))`. This will return records between and including the years provided.

geometry
string: WKT generated by `ala_geometry()`

caching
string: should the results be cached? Either "on" or "off"

Value

dataframe of matching species, with optional additional details.