Species lists

Retrieves a specified species list. See ala_lists to find all available species lists.

ala_list(druid, kvp = TRUE, verbose = ala_config()$verbose)

Arguments

druid
string: data resource UID of the list (i.e. the list identifier)

kvp
logical: include key-value pairs? Some lists contain information about the species in the form of key-value pairs

verbose
logical: show additional progress information?

Value

data.frame

Details

Note that this refers to pre-generated lists of species stored on the ALA servers. The similarly-named but different function specieslist provides a different function, namely listing the species matching a query or recorded as present in a search area.

References

https://lists.ala.org.au and the associated web services at https://lists.ala.org.au/ws

See also

Examples

if (FALSE) {
 all_lists <- ala_lists()
 ## find the "Field Guide apps species profiles" from Museum Victoria
 all_lists[grep("Field Guide", all_lists$listName), ]

 ## download the vertebrates one
 l <- ala_list(druid = "dr1146")
}