Find ALA species lists

Returns all available species lists. See ala_list to retrieve a specified list.

ala_lists(guid, offset = 0, max = 5000, verbose = ala_config()$verbose)

Arguments

guid
string: (optional) if provided, return only lists in which this GUID appears

offset
integer: the number of lists to skip. This supports paging

max
integer: the maximum number of lists to return. This supports paging

verbose
logical: show additional progress information?

Value

data.frame of list name and other details

References

https://lists.ala.org.au and the associated web services at https://lists.ala.org.au/ws

See also

Examples

if (FALSE) {
 ## lists that include the giant African snail Achatina fulica
 ##  (which is a notifiable pest species in some states)
 l <- ala_lists(search_guids("Achatina fulica")$guid)
}