Function to lookup species information from ALA, given names or unique identifiers.

Uses ALA name matching service. Species ids from this function can be used to search `ala_occurrences`

ala_taxa(
  term,
  term_type = "name",
  return_children = FALSE,
  include_counts = FALSE,
  caching = FALSE
)

Arguments

term
string, named list or dataframe: search term(s)

term_type
string: specifies which type of terms are provided in `term`. One of name `c('name', 'identifier')`

return_children
logical: return child concepts for the provided term(s)

include_counts
logical: return occurrence counts for all species returned

caching
logical: should the results be cached?

Value

dataframe of taxon information

Examples

if (FALSE) {
ala_taxa(term = "Reptilia")
ala_taxa(list(kingdom = "Plantae", genus = "Microseris"),
return_children = TRUE, include_counts = TRUE)
}