Check assertions in occurrences object

This provides a data.frame detailing the assertions that are found in a dataset returned from occurrences.

check_assertions(x)

Arguments

x
list: an object returned from occurrences

Value

A dataframe of assertions column names, descriptions and categories/error codes. If no assertions are in the dataset, NULL is returned.

References

http://api.ala.org.au/, http://biocache.ala.org.au/ws/assertions/codes

Examples

#download species data with all possible assertions

# NOT RUN {
 x <- occurrences(taxon="golden bowerbird",download_reason_id=10,qa=ala_fields("assertions")$name)
 asserts <- check_assertions(x) #data.frame of assertions, their description and column names
 asserts$description # List out descriptions of all (current) assertions

 tmp <- x$data[,names(x$data) %in% asserts$name] ## assertion columns from data
 which(colSums(tmp)>0) ## discard those not seen in the data
# }