Find images using occurrence ids

Find images using occurrence ids

occurrence_images(
  id,
  fq,
  download = FALSE,
  download_path,
  sounds = FALSE,
  verbose = ala_config()$verbose
)

Arguments

id
character: IDs of occurrences as single sring or vector of strings

fq
string: (optional) character string or vector of strings, specifying filters to be applied to the original query. These are of the form "INDEXEDFIELD:VALUE" e.g. "kingdom:Fungi".

download
logical: if TRUE download all images and add location to dataframe

download_path
string: (optional) filepath to download images to. If not given and download param is TRUE, will create an images folder

sounds
logical (optional) Image search also returns sound files. Ignored unless explicitly requested.

verbose
logical: show additional progress information? [default is set by ala_config()]

Value

Data frame of image results

References

Associated ALA web service for image search counts: https://images.ala.org.au/ws#/Search/search

Examples

if (FALSE) {
## Download all images for an occurrence with a CC BY-NC 4.0 licence
occurrence_image_search(id="d201f3e0-3e1d-47f1-94ce-9fc226cbc5ec",
fq="recognisedLicence:CC BY-NC 4.0",
download=TRUE)
}