Quick geographic plot of occurrence data

Generates a plot of occurrence data retrieved using occurrences.The plot uses an Australian basemap and colours the occurrence records dots according to parameters

occurrences_plot(x, filename = "Rplots.pdf", qa = c("fatal", "error"),
  grouped = FALSE, taxon_level = "species", pch, cex = 0.75, ...)

Arguments

x
list: a list object that has been downloaded using occurrences

filename
string: name of file to be created; defaults to RPlots.pdf

qa
string vector: list of record issues to be mapped; these can be assertion columnnames, or 'all' or 'none' or any combination of 'error', 'warning' or 'fatal'. Column or categories in your dataset can be viewed using check_assertions.

grouped
logical: TRUE creates a single plot for all observations; FALSE plots individual maps for the taxon level defined.

taxon_level
string: taxonomic level at which to create maps; possible values are 'species', 'genus', 'family' or 'order'

pch
single number or character representing point type. See description of pch in points.

cex
numeric: character (or symbol) expansion. See description of cex in points.

…
: other options passed to pdf()

Value

Generates a pdf that maps the distributions.

References

http://api.ala.org.au/

Examples

# NOT RUN {
#download some observations
x <- occurrences(taxon="Eucalyptus gunnii",download_reason_id=10)
occurrences_plot(x)
x <- occurrences(taxon="Cider Gum",download_reason_id=10)
occurrences_plot(x,"alaPlot.pdf",qa="fatal",grouped=FALSE, taxon_level="species",pch='+')
# }