Partial-name search

A partial-name search for species names & identifiers used at the ALA. If searching for a taxon name, and the scientific name or common name of the taxon are known, use search_names instead.

search_partial_name(taxon, geo_only = FALSE, output_format = "simple",
  index_type, limit)

Arguments

taxon
string: part of the scientific, common name of the taxa

geo_only
logical: if TRUE, only results that have geospatial occurrence records will be included

output_format
string: controls the print method for the returned object. Either "complete" (the complete data structure is displayed), or "simple" (a simplified version is displayed). Note that the complete data structure exists in both cases: this option only controls what is displayed when the object is printed to the console. The default output format is "simple"

index_type
string: the index type to limit. Values include: TAXON REGION COLLECTION INSTITUTION DATASET. By default, no index_type limit is applied

limit
numeric: the maximum number of matches returned (defaults to the server-side value - currently 10)

Value

A dataframe of results. The contents (column names) of the data frame will vary depending on the details of the search and the results.

References

Associated ALA web service: http://api.ala.org.au/#ws25

See also

Examples

# NOT RUN {
# find any names containing "allaba"
search_partial_name("allaba",output_format="simple")

# retrieve only species that have geolocated occurrence records
search_partial_name("Gallaba",geo_only=TRUE)
# }